[Refact] #54 동기적 실행이 필요한 곳 수정
initalSetting, setEmoji, addEmojiButtonEvent가 동기적으로 실행되도록 수정.
각 이모지에 대한 findAll이 최소 횟수로 실행되도록 수정.
