+++ js/init.js
@@ -47,14 +47,18 @@ function setSettingButtonsCss() {
 }
 
 export default function initialSetting() {
-  // 프로그램 첫 로드 시 Localstorage setting
-  if (localStorage.getItem('firstLoad') == null) {
-    localStorage.setItem('firstLoad', false);
+  return new Promise((resolve) => {
+    // 프로그램 첫 로드 시 Localstorage setting
+    if (localStorage.getItem('firstLoad') == null) {
+      localStorage.setItem('firstLoad', false);
 
-    localStorage.setItem('copy', 'auto');
-    localStorage.setItem('emojiSize', 'normal');
-    localStorage.setItem('recentNum', 20);
-    localStorage.setItem('windowState', 'popup');
-  }
-  setSettingButtonsCss();
+      localStorage.setItem('copy', 'auto');
+      localStorage.setItem('emojiSize', 'normal');
+      localStorage.setItem('recentNum', 20);
+      localStorage.setItem('windowState', 'popup');
+    }
+    setSettingButtonsCss();
+
+    resolve();
+  });
 }
